+++ src/main/java/com/spring/mummus/member/repository/MemberRepository.java
@@ -4,10 +4,8 @@ package com.spring.mummus.member.repository;
 import com.spring.mummus.member.domain.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Boolean existsByEmail(String email);
     Boolean existsByPhoneNumber(String phoneNumber);
-    Optional<Member> findByEmail(String email);
+    Member findByEmail(String email);
 }
