+++ src/main/java/com/spring/mummus/oauth2/service/OAuth2UserService.java
@@ -3,6 +3,8 @@ package com.spring.mummus.oauth2.service;
 import com.spring.mummus.member.domain.entity.Member;
 import com.spring.mummus.member.repository.MemberRepository;
 import com.spring.mummus.member.service.MemberService;
+import com.spring.mummus.oauth2.domain.KakaoOAuth2;
+import com.spring.mummus.oauth2.domain.OAuth2;
 import com.spring.mummus.oauth2.domain.PrincipalDetails;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,6 +15,7 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Component;
 
 import java.util.Map;
+import java.util.Objects;
 
 @Slf4j
 @Component
@@ -31,20 +34,25 @@ public class OAuth2UserService extends DefaultOAuth2UserService {
         log.info("userRequest getAdditionalParameters = {}", userRequest.getAdditionalParameters());
 
         OAuth2User oAuth2User = super.loadUser(userRequest);
+        Map<String, Object> attributes = oAuth2User.getAttributes();
         log.info("oAuth2User = " + oAuth2User);
-        log.info("oAuth2User getAttributes = " + oAuth2User.getAttributes());
+        log.info("attributes = " + attributes);
         log.info("oAuth2User getName = " + oAuth2User.getName());
         log.info("oAuth2User getAuthorities = " + oAuth2User.getAuthorities());
 
         String provider = userRequest.getClientRegistration().getRegistrationId();
-        Member member = null;
+        log.info("provider = {}", provider);
+        OAuth2 oAuth2 = null;
         switch (provider) {
-            case "kakao" -> member = memberService.singUp("d");
-            case "apple" -> member = memberService.singUp("d");
+            case "kakao" -> oAuth2 = new KakaoOAuth2(attributes);
         }
+
+        Member member = memberRepository.findByEmail(oAuth2.getEmail());
         log.info("member = " + member);
 
-        Map<String, Object> attributes = oAuth2User.getAttributes();
+        if (Objects.isNull(member)) {
+            memberService.singUp(oAuth2);
+        }
 
         return new PrincipalDetails(member, attributes);
     }
