+++ src/main/java/com/spring/mummus/pet/service/PetService.java
@@ -1,5 +1,6 @@
 package com.spring.mummus.pet.service;
 
+import com.spring.mummus.exception.exception.PetException;
 import com.spring.mummus.pet.domain.dto.RegisterPetRequest;
 import com.spring.mummus.pet.domain.entity.Pet;
 import com.spring.mummus.pet.repository.PetRepository;
@@ -7,17 +8,24 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import static com.spring.mummus.exception.enums.PetErrorCode.DUPLICATED_PET;
+
 @Service
 @RequiredArgsConstructor
 public class PetService {
-    private final PetReference petReference;
+//    private final PetReference petReference;
     private final PetRepository petRepository;
 
     @Transactional
     public Pet registerPet(RegisterPetRequest registerPetRequest) {
-        petReference.checkDuplicatedPet(registerPetRequest);
+        checkDuplicatedPet(registerPetRequest);
 
         return petRepository.save(registerPetRequest.toEntity());
     }
 
+    public void checkDuplicatedPet(RegisterPetRequest registerPetRequest) {
+        if (petRepository.existsByNameAndMemberId(registerPetRequest.getName(), registerPetRequest.getMemberId())) {
+            throw new PetException(DUPLICATED_PET);
+        }
+    }
 }
